+++ tests/test_tapinfo.py
@@ -1,9 +1,10 @@
 import unittest
+from unittest.mock import patch
 
 from skoolkittest import (SkoolKitTestCase, create_data_block,
                           create_tap_header_block, create_tap_data_block,
                           create_tzx_header_block, create_tzx_data_block)
-from skoolkit import SkoolKitError, get_word, VERSION
+from skoolkit import SkoolKitError, tapinfo, get_word, VERSION
 
 TZX_DATA_BLOCK = (16, 0, 0, 3, 0, 255, 0, 0)
 
@@ -16,6 +17,13 @@ TZX_DATA_BLOCK_DESC = """
 
 def _get_archive_info(text_id, text):
     return [text_id, len(text)] + [ord(c) for c in text]
+
+class MockBasicLister:
+    def list_basic(self, snapshot):
+        global mock_basic_lister
+        mock_basic_lister = self
+        self.snapshot = snapshot
+        return 'BASIC DONE!'
 
 class TapinfoTest(SkoolKitTestCase):
     def _write_tzx(self, blocks):
@@ -31,6 +39,11 @@ class TapinfoTest(SkoolKitTestCase):
         self.assertEqual(len(error), 0)
         final_block = TZX_DATA_BLOCK_DESC.format(blocks + 1).split('\n')
         self.assertEqual(['Version: 1.20'] + exp_output + final_block, output)
+
+    def _test_bad_spec(self, option, bad_spec, exp_error):
+        with self.assertRaises(SkoolKitError) as cm:
+            self.run_tapinfo('{} {} test.tap'.format(option, bad_spec))
+        self.assertEqual(cm.exception.args[0], '{}: {}'.format(exp_error, bad_spec))
 
     def test_no_arguments(self):
         output, error = self.run_tapinfo(catch_exit=2)
@@ -476,6 +489,38 @@ class TapinfoTest(SkoolKitTestCase):
         output, error = self.run_tapinfo('-b 2z {}'.format(tzxfile))
         self.assertEqual(len(error), 0)
         self.assertEqual(exp_output, output)
+
+    @patch.object(tapinfo, 'BasicLister', MockBasicLister)
+    def test_option_B_tap(self):
+        prog = [10] * 10
+        tap_data = create_tap_data_block(prog)
+        tapfile = self.write_bin_file(tap_data, suffix='.tap')
+        exp_snapshot = [0] * 23755 + prog
+        output, error = self.run_tapinfo('-B 1 {}'.format(tapfile))
+        self.assertEqual(error, '')
+        self.assertEqual(['BASIC DONE!'], output)
+        self.assertEqual(exp_snapshot, mock_basic_lister.snapshot)
+
+    @patch.object(tapinfo, 'BasicLister', MockBasicLister)
+    def test_option_basic_tzx_with_address(self):
+        prefix = [1] * 5
+        address = 23755 - len(prefix)
+        prog = [12] * 12
+        data = prefix + prog
+        blocks = [create_tzx_data_block(data)]
+        tzxfile = self._write_tzx(blocks)
+        exp_snapshot = [0] * address + data
+        output, error = self.run_tapinfo('--basic 1,{} {}'.format(address, tzxfile))
+        self.assertEqual(error, '')
+        self.assertEqual(['BASIC DONE!'], output)
+        self.assertEqual(exp_snapshot, mock_basic_lister.snapshot)
+
+    def test_option_B_with_invalid_block_spec(self):
+        exp_error = 'Invalid block specification'
+        self._test_bad_spec('-B', 'q', exp_error)
+        self._test_bad_spec('--basic', '1,z', exp_error)
+        self._test_bad_spec('-B', '1,2,3', exp_error)
+        self._test_bad_spec('--basic', '?,+', exp_error)
 
     def test_option_V(self):
         for option in ('-V', '--version'):